update the loading of category icon in task card and add the mapping of resDrawables in presentation
