+++ app/src/main/java/com/example/tudeeapp/presentation/screen/home/state/TaskUiState.kt
@@ -1,6 +1,5 @@
 package com.example.tudeeapp.presentation.screen.home.state
 
-import com.example.tudeeapp.R
 import com.example.tudeeapp.domain.models.TaskPriority
 import com.example.tudeeapp.domain.models.TaskStatus
 
@@ -12,5 +11,6 @@ data class TaskUiState(
     val priority: TaskPriority,
     val priorityResIcon: Int,
     val status: TaskStatus,
-    val categoryIcon: Int=R.drawable.ic_loading
+    val categoryIcon: String = "",
+    val isCategoryPredefined: Boolean = false
 )
