+++ app/src/main/java/com/example/tudeeapp/presentation/screen/taskManagement/components/CategoryGrid.kt
@@ -16,11 +16,11 @@ import androidx.compose.ui.unit.dp
 import androidx.core.net.toUri
 import coil3.compose.rememberAsyncImagePainter
 import com.example.tudeeapp.R
-import com.example.tudeeapp.data.mapper.DataConstant.toResDrawables
 import com.example.tudeeapp.presentation.screen.taskManagement.CategoryUiState
 import com.example.tudeeapp.presentation.common.components.CategoryItem
 import com.example.tudeeapp.presentation.common.extentions.BasePreview
 import com.example.tudeeapp.presentation.design_system.theme.Theme
+import com.example.tudeeapp.presentation.mapper.toResDrawables
 
 @Composable
 fun CategoryGrid(
